pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "hxse", email = "narutohy@163.com"},
]
dependencies = [
    "requests==2.25.1",
    "beautifulsoup4==4.9.3",
    "m3u8==0.8.0",
    "pycryptodome",
    "cloudscraper==1.2.58",
    "python-ffmpeg>=1.0.14",
    "fire>=0.4.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[tool.pdm]
[[tool.pdm.source]]
name = "pypi"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
verify_ssl = true

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"
